+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/button/DefaultCheckBoxBuilder.java
@@ -18,11 +18,11 @@
  */
 package is.codion.swing.common.ui.component.button;
 
-import is.codion.swing.common.model.component.button.NullableToggleButtonModel;
-
 import javax.swing.JCheckBox;
 import javax.swing.JToggleButton;
 import javax.swing.SwingConstants;
+
+import static is.codion.swing.common.model.component.button.NullableToggleButtonModel.nullableToggleButtonModel;
 
 final class DefaultCheckBoxBuilder extends DefaultToggleButtonBuilder<JCheckBox, CheckBoxBuilder> implements CheckBoxBuilder {
 
@@ -40,6 +40,6 @@ final class DefaultCheckBoxBuilder extends DefaultToggleButtonBuilder<JCheckBox,
 
 	@Override
 	protected JToggleButton createToggleButton() {
-		return nullable ? new NullableCheckBox(new NullableToggleButtonModel()) : new JCheckBox();
+		return nullable ? new NullableCheckBox(nullableToggleButtonModel()) : new JCheckBox();
 	}
 }